lugx-microservices/game-service/requirements.txt
flask
supabase==2.15.3
python-dotenv
prometheus_client